🧪 test: move params tests to integration tests
